BE-2288: Separate dynamodb related methods in entc/gen pkg
